Add documentation including rustdoc, jsdoc, & clap cli docs to server
