+++ .circleci/config.yml
@@ -186,8 +186,8 @@ jobs:
             cargo install --path snarkOS --root . --locked
             rm -rf snarkOS
             ./bin/snarkos start --dev 0 --beacon APrivateKey1zkp8CZNn3yeCseEtxuVPbDCwSyhGW6yZKUYKfgXmcpoGPWH --nodisplay 1> /dev/null &
-            cargo install --path rust/rest --root . --locked
-            ./bin/aleo-develop start --uri 0.0.0.0:4321 --peer http://localhost:3030 1> /dev/null &
+            cargo install --path rust/develop --root . --locked
+            ./bin/aleo-develop start --server-address 0.0.0.0:4040 --peer http://localhost:3030 -d 1> /dev/null &
             sleep 15
             cd rust
             cargo test test_transfer -- --ignored --nocapture --test-threads=1
