+++ Cargo.toml
@@ -18,7 +18,7 @@ license = "GPL-3.0"
 edition = "2021"
 
 [workspace]
-members = [ "rust", "wasm", "rust/rest" ]
+members = [ "rust", "wasm", "rust/develop"]
 
 [workspace.dependencies.snarkvm]
 version = "0.9.15"
