+++ rust/develop/src/helpers/auth.rs
