--- rust/rest/README.md
@@ -1,7 +1,0 @@
-# snarkos-node-rest
-
-[![Crates.io](https://img.shields.io/crates/v/snarkos-node-rest.svg?color=neon)](https://crates.io/crates/snarkos-node-rest)
-[![Authors](https://img.shields.io/badge/authors-Aleo-orange.svg)](https://aleo.org)
-[![License](https://img.shields.io/badge/License-GPLv3-blue.svg)](./LICENSE.md)
-
-The `snarkos-node-rest` crate provides a REST API for the `snarkos` node.
