+++ sdk/tests/development.integration.ts
@@ -9,7 +9,7 @@ describe('DevelopmentServer', () => {
     let client: DevelopmentClient;
 
     beforeEach(() => {
-        client = new DevelopmentClient("http://0.0.0.0:4321");
+        client = new DevelopmentClient("http://0.0.0.0:4040");
     });
 
     describe('Deploy & Execute', () => {
@@ -35,7 +35,9 @@ describe('DevelopmentServer', () => {
             let deploy_transaction_id = "";
             for (let i = 0; i < 4; i++) {
                 try {
+                    console.log("Trying to deploy program");
                     deploy_transaction_id = await client.deployProgram(helloProgram, 6000000, privateKeyString);
+                    await wait(5000);
                     break;
                 }
                 catch (e) {
@@ -43,15 +45,17 @@ describe('DevelopmentServer', () => {
             }
             // If the transaction failed above, try one more time
             if (deploy_transaction_id === "") {
+                console.log("Trying a final time to deploy program")
                 deploy_transaction_id = await client.deployProgram(helloProgram, 6000000, privateKeyString);
             }
             expect(deploy_transaction_id).toBeTruthy();
 
-            await wait(10000);
+            await wait(20000);
 
             let execute_transaction_id = "";
             for (let i = 0; i < 3; i++) {
                 try {
+                    console.log("Trying a final time to to execute program");
                     execute_transaction_id = await client.executeProgram(helloProgramId, "main", 0, ["5u32", "5u32"], privateKeyString);
                     expect(execute_transaction_id).toBeTruthy();
                     break;
@@ -61,6 +65,7 @@ describe('DevelopmentServer', () => {
 
             // If the transaction failed above, try one more time
             if (execute_transaction_id === "") {
+                console.log("Trying to execute one more time");
                 execute_transaction_id = await client.deployProgram(helloProgram, 6000000, privateKeyString);
             }
             expect(execute_transaction_id).toBeTruthy();
